refactor: consolidate backend into chatagent package
