+++ backend/chatagent/models/world_state.py
