+++ backend/chatagent/schemas/agent.schema.json
