+++ backend/chatagent/schemas/event.schema.json
