+++ backend/chatagent/schemas/simulation_config.schema.json
