+++ backend/chatagent/schemas/world_state.schema.json
