+++ backend/chatagent/services/world.py
