+++ backend/chatagent/templates/index.html
@@ -21,6 +21,6 @@
   </div>
   <div>Tick: <span id="tick">0</span></div>
   <pre id="snapshot"></pre>
-  <script src="/static/app.js"></script>
+  <script src="/static/chatagent.js"></script>
 </body>
 </html>
