+++ backend/tests/test_scenario_validation.py
@@ -2,7 +2,10 @@ from pathlib import Path
 
 import pytest
 
-from app.services.validation import validate_scenario_file, ScenarioValidationError
+from chatagent.services.validation import (
+    validate_scenario_file,
+    ScenarioValidationError,
+)
 
 
 DATA_DIR = Path(__file__).resolve().parents[1] / "data"
